requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
python-multipart==0.0.9
